Fixed QuantityType in Edit and made a composite POJO for list items.
Added a QuantityTypeAdapter to the Edit activity to handle the quantity
type dropdown.
Created a composite POJO for list items which have a listItem,
quantityType, and a category. This allows for 1:1 relations in tables
and a class which combines them all. Room does not do this by default
(which has to do with lazy loading).

Added a method to the ListItemDao to get all the fields needed
populating all objects in the composite list item POJO.

Altered code everywhere to allow the composite list item object to be
passed around in intents and used in the main activity listing.